[primsTS][fmha][context] Add differing QK/PV dtype support to primsTS context attention
Split the single q_dtype field into qk_dtype and pv_dtype across the context geometry, compile-spec, and validation paths, and thread both through to FmhaTs as separate in_qk_dtype/in_pv_dtype parameters. Q and K must still share one dtype, but V may now use a different one (e.g. QK-BF16/PV-FP8). Relax dtype validation and update tests accordingly.

+++ tests/attention/test_attention_ts_context.py
@@ -813,7 +813,8 @@ def test_attention_ts_context_d256_pipeline_policy_is_semantic_and_capacity_safe
     """Every D256 topology provides enough stages for its K/V cadence."""
 
     kernel = FmhaTs(
-        in_dtype=input_dtype,
+        in_qk_dtype=input_dtype,
+        in_pv_dtype=input_dtype,
         out_dtype=output_dtype,
         d=256,
         is_persistent=True,
@@ -865,7 +866,8 @@ def test_attention_ts_context_page_window_fits_static_geometry_and_capacity(
     """The wider page-ID handoff is selected only when its SMEM ring fits."""
 
     cfg = FmhaTs(
-        in_dtype=input_dtype,
+        in_qk_dtype=input_dtype,
+        in_pv_dtype=input_dtype,
         out_dtype=BFloat16,
         d=256,
         is_persistent=True,
@@ -1050,7 +1052,8 @@ def test_attention_ts_context_d128_paged_clc_task_graph_is_safe(
     """The paired D128 CLC graph is valid without a page-ID ring."""
 
     kernel = FmhaTs(
-        in_dtype=input_dtype,
+        in_qk_dtype=input_dtype,
+        in_pv_dtype=input_dtype,
         out_dtype=Float16,
         d=128,
         is_persistent=True,
@@ -1106,7 +1109,8 @@ def test_attention_ts_context_d256_live_paged_clc_uses_distinct_auxiliary_warps(
 ):
     """The live-ragged D256 CLC and page-ID producers cannot overlap."""
     kernel = FmhaTs(
-        in_dtype=input_dtype,
+        in_qk_dtype=input_dtype,
+        in_pv_dtype=input_dtype,
         out_dtype=Float16,
         d=256,
         is_persistent=True,
@@ -1185,7 +1189,8 @@ def test_attention_ts_context_d256_uniform_paged_static_scheduler_is_safe(
 
     balance_causal_workload = not has_q_offset
     kernel = FmhaTs(
-        in_dtype=input_dtype,
+        in_qk_dtype=input_dtype,
+        in_pv_dtype=input_dtype,
         out_dtype=Float16,
         d=256,
         is_persistent=True,
